Deletes most of the old files that were not in use
