+++ src/main/java/edu/ncsu/dlf/servlet/LoginServlet.java
@@ -22,38 +22,49 @@ import org.json.JSONObject;
 public class LoginServlet extends HttpServlet {
 	private static final long serialVersionUID = 1L;
 	@Override
-	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+	public void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+
+		//So if someone hits /login2 without going through our flow,
+		//they will just see the client_id displayed on the page
+		//Display error page instead?
 		if (req.getParameter("code") == null) {
 		    resp.setContentType("application/json");
 		    JSONObject jobj = new JSONObject();
 		    try {
 		        jobj.put("client_id", System.getenv("GITHUB_ID"));
-	            jobj.write(resp.getWriter()); 
+	          jobj.write(resp.getWriter());
 		    } catch (JSONException e) {
 		        e.printStackTrace();
 		        resp.sendError(500);
 		    }
 		    return;
 		}
-	    
-	    
-	    HttpPost request = null;
+
+		//Make a POST to the GitHub OAuth API, with the code received from the previous page
+	  HttpPost request = null;
 		try {
 			URIBuilder builder = new URIBuilder("https://github.com/login/oauth/access_token");
 			builder.addParameter("client_id", System.getenv("GITHUB_ID"));
 			builder.addParameter("client_secret", System.getenv("GITHUB_API"));
 			builder.addParameter("code", req.getParameter("code"));
-			
+
 			request = new HttpPost(builder.build());
-			
 			request.setHeader("accept", "application/json");
-	
-			
+
 			HttpClient client = HttpClients.createDefault();
-			HttpResponse response = client.execute(request);
-			String body = HttpUtils.getResponseBody(response);
-			resp.setContentType("application/json");
-			resp.getWriter().write(body);
+			HttpResponse authResponse = client.execute(request);
+
+			String accessToken = "";
+			//Extracts the access_token from the response
+			try {
+				JSONObject responseJSON = new JSONObject(HttpUtils.getResponseBody(authResponse));
+				accessToken = (String) responseJSON.get("access_token");
+			} catch (JSONException e) {
+				e.printStackTrace();
+			}
+
+			resp.sendRedirect(req.getContextPath() + "/tool?access_token=" + accessToken);
+
 		} catch(URISyntaxException e) {
 			e.printStackTrace();
 		} finally{
@@ -62,5 +73,5 @@ public class LoginServlet extends HttpServlet {
 			}
 		}
 	}
-	
+
 }
