--- src/main/java/edu/ncsu/dlf/utils/JSONUtils.java
@@ -1,22 +1,0 @@
-package edu.ncsu.dlf.utils;
-
-import java.util.List;
-
-import edu.ncsu.dlf.model.Review;
-
-import org.json.JSONArray;
-import org.json.JSONException;
-
-public final class JSONUtils {
-
-    private JSONUtils() {}
-
-    public static JSONArray toJSON(List<Review> list) throws JSONException {
-        JSONArray jarr = new JSONArray();
-        for(Review r: list) {
-            jarr.put(r.toJSON());
-        }
-        return jarr;
-    }
-    
-}
