--- src/main/webapp/index2.html
@@ -1,38 +1,0 @@
-<html>
-<head>
-  <title>PDF Reviewer</title>
-
-  <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons">
-  <link rel="stylesheet" href="https://code.getmdl.io/1.3.0/material.indigo-pink.min.css">
-  <script defer src="https://code.getmdl.io/1.3.0/material.min.js"></script>
-  <link rel="icon" href="favicon.png" type="image/png">
-</head>
-<body>
-  <style>
-    body {
-      text-align: center;
-    }
-  .mdl-button {
-    margin-top: 40px;
-    font-size: larger;
-    text-align: center;
-    color: white;
-    width: 400px;
-    height:100px;
-    background-color: rgb(7, 157, 184);
-  }
-  </style>
-
-  <button id="login" class="mdl-button mdl-js-button mdl-button--raised mdl-js-ripple-effect ">
-    <img src="images/GitHub-Mark-Light-64px.png">  Login with GitHub
-  </button>
-  <!--<button id="login" class="btn btn-primary btn-lg">
-    <img src="images/github-icon.png"> Login with GitHub
-  </button>-->
-
-
-  <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
-  <script src="./js/login2.js"></script>
-
-</body>
-</html>
